+++ agentpy/__init__.py
@@ -10,7 +10,7 @@ Examples: https://agentpy.readthedocs.io/en/latest/model_library.html
 __all__ = [
     '__version__',
     'Model', 'Environment', 'Network', 'Grid',
-    'Agent', 'AgentList', 'EnvDict',
+    'Agent', 'AgentList', 'AttrList', 'EnvDict',
     'Experiment',
     'DataDict', 'load',
     'sample', 'sample_discrete', 'sample_saltelli',
@@ -21,7 +21,7 @@ __version__ = "0.0.5.dev"
 
 # Objects
 from .framework import Model, Environment, Network, Grid
-from .framework import Agent, AgentList, EnvDict
+from .framework import Agent, AgentList, AttrList, EnvDict
 from .experiment import Experiment
 from .output import DataDict, load
 from .sampling import sample, sample_discrete, sample_saltelli